requirements.txt
reportlab
cairosvg
python-chess
pdf2image
Pillow
cairocffi
tinycss2
lxml
